+++ src/common/utils/index.js
@@ -1,4 +1,5 @@
 export * from './calculateTotalAmount';
+export * from './fetchInvoiceLines';
 export * from './getAccountingCodeOptions';
 export * from './getLegacyTokenHeader';
 export * from './getOptions';