+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 7.1.0 (IN PROGRESS)
 
+* Fetch all invoice lines per invoice on export. Refs UINV-572.
+
 ## [7.0.0](https://github.com/folio-org/ui-invoice/tree/v7.0.0) (2025-03-13)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v6.1.2...v7.0.0)